platform: qpnp-power-on: Simulate press event
Status of the keys connected to the KPDPWR and RESIN lines
are identified by reading corresponding bits in RT status
register. If the state has been changed by the time interrupt
is delivered then the press event will be missed.

Maintain a state variable to find unbalanced release events
and simulate a press event.

Change-Id: I809c8166796e2436dbd203e4aa0c304d3d612226
Signed-off-by: Mohan Pallaka <mpallaka@codeaurora.org>

Conflicts:
	drivers/platform/msm/qpnp-power-on.c