msm: mdss: dsi: add support to configure regulator low power mode
For idle screen usecases in ultra-low power mode, it may be possible to
reduce the requested load on panel and controller regulators. Add
support for this by introducing a new regulator load entry and configure
the new load whenever DSI core power is turned off in low power mode.

Change-Id: I52f78d04f877b7095e552598ab997d977acabf02
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Zachariah <zkennedy87@gmail.com>

Conflicts:
	drivers/video/msm/mdss/mdss_dsi.h